requirements.txt
selenium>=4.0.0
beautifulsoup4>=4.9.3
pandas>=1.3.0
scikit-learn>=1.0.0
matplotlib>=3.3.0
numpy>=1.21.0
webdriver-manager>=4.0.0
bertopic>=0.15.0
sentence-transformers>=2.2.0
umap-learn>=0.5.3
